requirements.txt
flask
werkzeug
pillow
numpy
opencv-python-headless
gunicorn
requests


